bugged strategy builder

+++ src/main/java/org/tradingsignal/pojo/yahoo/Result.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class Result {
     private Meta meta;
-    private List<Integer> timestamp;
+    private List<Long> timestamp;
     private Events events;
     private Indicators indicators;
 }
